requirements.txt
asyncio==3.4.3
discord==1.7.3
python-dotenv==0.18.0
youtube-dl==2021.12.17